+++ ebpf/aya-bpf/src/maps/lpm_trie.rs
@@ -1,6 +1,6 @@
 use core::{cell::UnsafeCell, marker::PhantomData, mem, ptr::NonNull};
 
-use aya_bpf_bindings::bindings::BPF_F_NO_PREALLOC;
+use aya_ebpf_bindings::bindings::BPF_F_NO_PREALLOC;
 use aya_ebpf_cty::{c_long, c_void};
 
 use crate::{