+++ ebpf/aya-bpf/src/maps/xdp/dev_map_hash.rs
@@ -1,6 +1,6 @@
 use core::{cell::UnsafeCell, mem, num::NonZeroU32, ptr::NonNull};
 
-use aya_bpf_bindings::bindings::bpf_devmap_val;
+use aya_ebpf_bindings::bindings::bpf_devmap_val;
 use aya_ebpf_cty::c_void;
 
 use super::{dev_map::DevMapValue, try_redirect_map};