+++ ebpf/aya-bpf/src/maps/xdp/xsk_map.rs
@@ -1,6 +1,6 @@
 use core::{cell::UnsafeCell, mem, ptr::NonNull};
 
-use aya_bpf_bindings::bindings::bpf_xdp_sock;
+use aya_ebpf_bindings::bindings::bpf_xdp_sock;
 use aya_ebpf_cty::c_void;
 
 use super::try_redirect_map;